Changed the int variables to doubles so that we wouldn't lose value. Reported in https://app.asana.com/0/9404392569325/9404392569327

+++ udemy_c-sharp-beginner/Program.cs
@@ -9,7 +9,7 @@ namespace udemy_c_sharp_beginner
     {
         static void Main(string[] args)
         {
-            int num1, num2;
+            double num1, num2;
             num1 = 24;
             num2 = 13;
             Console.WriteLine(num1 + num2);
